+++ packages/cli/__tests__/zosjobs/__system__/search/__scripts__/job/missing_search_string.sh
@@ -1,11 +1,8 @@
 #!/bin/bash
 
-zowe jobs search job "$1" $2
+zowe jobs search job "$1"
 RC=$?
-if [ $RC -gt 0 ]
-then
-    echo $STATUS 1>&2
+if [ $RC -gt 0 ]; then
     echo "The search spool job command returned a non-zero rc: $RC" 1>&2
     exit $RC
 fi
-