+++ packages/workflows/__tests__/__system__/Cancel.system.test.ts
@@ -11,14 +11,14 @@
 
 import { CancelWorkflow, CreateWorkflow, DeleteWorkflow } from "../../src";
 import { Imperative, ImperativeError, Session } from "@zowe/imperative";
-import { ITestEnvironment } from "@zowe/cli-test-utils";
 import { TestEnvironment } from "../../../../__tests__/__src__/environment/TestEnvironment";
 import { ITestPropertiesSchema } from "../../../../__tests__/__src__/properties/ITestPropertiesSchema";
 import { inspect } from "util";
 import { getUniqueDatasetName } from "../../../../__tests__/__src__/TestUtils";
 import { noWorkflowKey, WrongWorkflowKey } from "../../src/WorkflowConstants";
-import { Upload, ZosFilesConstants } from "@zowe/zos-files-for-zowe-sdk";
-import { ZosmfRestClient, nozOSMFVersion, noSession } from "@zowe/core-for-zowe-sdk";
+import { Upload } from "@zowe/zos-files-for-zowe-sdk";
+import { nozOSMFVersion, noSession } from "@zowe/core-for-zowe-sdk";
+import { ITestEnvironment } from "../../../../__tests__/__src__/environment/ITestEnvironment";
 
 let REAL_SESSION: Session;
 let testEnvironment: ITestEnvironment<ITestPropertiesSchema>;
@@ -70,19 +70,7 @@ describe("Cancel workflow", () => {
         beforeAll(async () => {
             // Upload files only for successful scenarios
             await Upload.fileToUssFile(REAL_SESSION, workflow, definitionFile, { binary: true });
-        });
-        afterAll(async () => {
-            let error;
-            let response;
-
-            const endpoint: string = ZosFilesConstants.RESOURCE + ZosFilesConstants.RES_USS_FILES;
-            // Deleting uploaded workflow file
-            try {
-                const wfEndpoint = endpoint + definitionFile;
-                response = await ZosmfRestClient.deleteExpectString(REAL_SESSION, wfEndpoint);
-            } catch (err) {
-                error = err;
-            }
+            testEnvironment.resources.files.push(definitionFile);
         });
         beforeEach(async () => {
             const response = await CreateWorkflow.createWorkflow(REAL_SESSION, wfName, definitionFile, system, owner);